Modification to support PenMount 1610
Modification to input-raw.c to support PenMount 1610 as issue described in 
#21

The touch signals is listed in the issue. Changes:

Touch action is ended when within a SYN_REPORT there is no ABS_X or ABS_Y and the pen down signal (BTN_TOUCH BTN_LEFT , value == 1) is not called
EV_ABS ABS_MT_WIDTH_MINOR is added as a pressure signal.